refactor synchronization
